requirements.txt
fastmcp
google-genai
python-dotenv
biopython>=1.80
pytest>=7.0
